refactor[#226] : Modal UI 카멜 케이스 + 모달명 변경
