+++ src/features/setting/ui/passwordCheckModal.stories.tsx
@@ -1,0 +1,13 @@
+import { StoryObj } from "@storybook/react";
+import { PasswordCheckModal } from "./passwordCheckModal";
+
+export default {
+  title: "features/setting/CancellationCheckModal",
+  component: PasswordCheckModal,
+};
+
+type Story = StoryObj<typeof PasswordCheckModal>;
+
+export const Default: Story = {
+  render: () => <PasswordCheckModal onClose={async () => {}} />,
+};
